remove: home
ホームディレクトリをhomeからtodoに変更しました
